% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_indicador}
\alias{calc_indicador}
\title{Wrapper para as funções de cálculo de indicador}
\usage{
calc_indicador(indicador, arg1, arg2, arg3)
}
\arguments{
\item{indicador}{Identificação do indicador a ser calculado, e.g., "1A", "17", etc}

\item{arg1}{Primeiro argumento do indicador}

\item{arg2}{Segundo argumento do indicador}

\item{arg3}{Terceiro argumento do indicador}
}
\value{
Valor do indicador
}
\description{
Wrapper para as funções de cálculo de indicador
}
