% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_10.R
\name{calc_indicador_10}
\alias{calc_indicador_10}
\title{Calcula o indicador 10: "Percentual de matrículas de educação de jovens e adultos na forma integrada à educação profissional"}
\usage{
calc_indicador_10(df_matricula, ano, verbose = TRUE)
}
\description{
Calcula o indicador 10: "Percentual de matrículas de educação de jovens e adultos na forma integrada à educação profissional"
}
