% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_11.R
\name{calc_indicador_11}
\alias{calc_indicador_11}
\title{Calcula o indicador 11A: "Matrículas em educação profissional técnica de nível médio"}
\usage{
calc_indicador_11(df_matricula, ano, verbose = TRUE)
}
\description{
Calcula o indicador 11A: "Matrículas em educação profissional técnica de nível médio"
}
