% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_11.R
\name{calc_indicador_11A}
\alias{calc_indicador_11A}
\title{Calcula o indicador 11A: "Matrículas em educação profissional técnica de nível médio"}
\usage{
calc_indicador_11A(df_matricula)
}
\arguments{
\item{df_matricula}{DataFrame com dados carregados da tabela "matricula" do Censo Escolar}
}
\value{
Indicador 11A
}
\description{
Calcula o indicador 11A: "Matrículas em educação profissional técnica de nível médio"
}
