% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_12.R
\name{calc_indicador_12B}
\alias{calc_indicador_12B}
\title{Calcula o indicador 12B: "Taxa de escolarização líquida ajustada na educação superior da população de 18 a 24 anos"}
\usage{
calc_indicador_12B(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 12B em porcentagem
}
\description{
Calcula o indicador 12B: "Taxa de escolarização líquida ajustada na educação superior da população de 18 a 24 anos"
}
