% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_13.R
\name{calc_indicador_13A}
\alias{calc_indicador_13A}
\title{Calcula o indicador 13A: "Percentual de funções docentes na educação superior com mestrado e doutorado"}
\usage{
calc_indicador_13A(df_docentes, df_ies, periodo)
}
\arguments{
\item{df_docentes}{DataFrame com dados carregados da tabela "docentes" do Censo da Educação Superior}

\item{df_ies}{DataFrame com dados carregados da tabela "ies" do Censo da Educação Superior}

\item{periodo}{Ano do censo}
}
\value{
Indicador 13A em porcentagem
}
\description{
Calcula o indicador 13A: "Percentual de funções docentes na educação superior com mestrado e doutorado"
}
