% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_16.R
\name{calc_indicador_16}
\alias{calc_indicador_16}
\title{Calcula o indicador 16: "Percentual de professores da educação básica com pós-graduação lato sensu ou stricto senso"}
\usage{
calc_indicador_16(df_docentes)
}
\arguments{
\item{df_docentes}{DataFrame com dados carregados da tabela "docentes" do Censo Escolar}
}
\value{
Indicador 16 em porcentagem
}
\description{
Calcula o indicador 16: "Percentual de professores da educação básica com pós-graduação lato sensu ou stricto senso"
}
