% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_17.R
\name{calc_indicador_17_anual}
\alias{calc_indicador_17_anual}
\title{Calcula o indicador 17: "Razão entre salários dos professores da educação básica, na rede pública (não federal), e não professores, com escolaridade equivalente"}
\usage{
calc_indicador_17_anual(df_anual, verbose = TRUE)
}
\arguments{
\item{df_anual}{DataFrame com dados carregados da PNAD Contínua anual}

\item{verbose}{exibe informações no console se True}
}
\value{
Indicador 17 em porcentagem
}
\description{
Calcula o indicador 17: "Razão entre salários dos professores da educação básica, na rede pública (não federal), e não professores, com escolaridade equivalente"
}
