% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_1.R
\name{calc_indicador_1A}
\alias{calc_indicador_1A}
\title{Calcula o indicador 1A: "Percentual de crianças de 4 a 5 anos na pré-escola"}
\usage{
calc_indicador_1A(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 1A em porcentagem
}
\description{
Calcula o indicador 1A: "Percentual de crianças de 4 a 5 anos na pré-escola"
}
