% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_1.R
\name{calc_indicador_1A_censo}
\alias{calc_indicador_1A_censo}
\title{Calcula o indicador 1A: "Percentual de crianças de 4 a 5 anos na pre-escola"}
\usage{
calc_indicador_1A_censo(df_mat, df_pop, ano, verbose = TRUE)
}
\arguments{
\item{df_mat}{tabela matrícula do censo escolar (a partir de 2014)}

\item{df_pop}{tabela com total da população fornecido pela prefeitura}

\item{ano}{ano do censo fornecido (>= 2014)}

\item{verbose}{be communicative?}
}
\description{
Calcula o indicador 1A: "Percentual de crianças de 4 a 5 anos na pre-escola"
}
