% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_1.R
\name{calc_indicador_1B_censo}
\alias{calc_indicador_1B_censo}
\title{Calcula o indicador 1B: "Percentual da população de 0 a 3 anos que frequenta a escola"}
\usage{
calc_indicador_1B_censo(df_mat, df_pop, ano, verbose = TRUE)
}
\arguments{
\item{df_mat}{tabela matrícula do censo escolar (a partir de 2014)}

\item{df_pop}{tabela com total da população fornecido pela prefeitura}
}
\description{
Calcula o indicador 1B: "Percentual da população de 0 a 3 anos que frequenta a escola"
}
