% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_2.R
\name{calc_indicador_2A}
\alias{calc_indicador_2A}
\title{Calcula o indicador 2A: "Percentual a população de 6 a 14 anos que frequenta a escola"}
\usage{
calc_indicador_2A(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 2A em porcentagem
}
\description{
Calcula o indicador 2A: "Percentual a população de 6 a 14 anos que frequenta a escola"
}
