% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_2.R
\name{calc_indicador_2B}
\alias{calc_indicador_2B}
\title{Calcula o indicador 2B: "Percentual de pessoas de 16 anos com pelo menos o ensino fundamental concluído"}
\usage{
calc_indicador_2B(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 2B em porcentagem
}
\description{
Calcula o indicador 2B: "Percentual de pessoas de 16 anos com pelo menos o ensino fundamental concluído"
}
