% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_3.R
\name{calc_indicador_3A}
\alias{calc_indicador_3A}
\title{Calcula o indicador 3A: "Percentual da população de 15 a 17 anos que frequenta a escola"}
\usage{
calc_indicador_3A(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 3A em porcentagem
}
\description{
Calcula o indicador 3A: "Percentual da população de 15 a 17 anos que frequenta a escola"
}
