% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_3.R
\name{calc_indicador_3B_censo}
\alias{calc_indicador_3B_censo}
\title{Calcula o indicador 3B: "Taxa líquida de matrícula no Ensino Médio"}
\usage{
calc_indicador_3B_censo(df_mat, df_pop, ano, verbose = TRUE)
}
\arguments{
\item{df_mat}{tabela matrícula do censo escolar (a partir de 2014)}

\item{df_pop}{tabela com total da população fornecido pela prefeitura}

\item{ano}{ano do censo fornecido (>= 2014)}

\item{verbose}{be communicative?}
}
\description{
Calcula o indicador 3B: "Taxa líquida de matrícula no Ensino Médio"
}
