% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_5.R
\name{calc_indicador_5A}
\alias{calc_indicador_5A}
\title{Calcula o indicador 5A: "Proficiência dos alunos do 3o ano do ensino fundamental em Leitura"}
\usage{
calc_indicador_5A(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da Avaliação Nacional da Alfabetização}
}
\value{
Indicador 5A em porcentagem
}
\description{
Calcula o indicador 5A: "Proficiência dos alunos do 3o ano do ensino fundamental em Leitura"
}
