% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_5.R
\name{calc_indicador_5B}
\alias{calc_indicador_5B}
\title{Calcula o indicador 5B: "Proficiência dos alunos do 3o ano do ensino fundamental em Escrita"}
\usage{
calc_indicador_5B(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da Avaliação Nacional da Alfabetização}
}
\value{
Indicador 5B em porcentagem
}
\description{
Calcula o indicador 5B: "Proficiência dos alunos do 3o ano do ensino fundamental em Escrita"
}
