% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_5.R
\name{calc_indicador_5C}
\alias{calc_indicador_5C}
\title{Calcula o indicador 5C: "Taxa de estudantes com proficiência insuficiente em Matemática"}
\usage{
calc_indicador_5C(df, verbose = TRUE)
}
\arguments{
\item{df}{DataFrame com dados carregados da Avaliação Nacional da Alfabetização}

\item{verbose}{exibe informações no console se True}
}
\value{
Indicador 5C em porcentagem
}
\description{
Calcula o indicador 5C: "Taxa de estudantes com proficiência insuficiente em Matemática"
}
