% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_6.R
\name{calc_indicador_6A}
\alias{calc_indicador_6A}
\title{Calcula o indicador 6A: "Percentual de escolas públicas com estudantes que permanecem pelo menos 7h em atividades escolares"}
\usage{
calc_indicador_6A(df_matricula, ano, verbose = TRUE)
}
\description{
Calcula o indicador 6A: "Percentual de escolas públicas com estudantes que permanecem pelo menos 7h em atividades escolares"
}
