% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_6.R
\name{calc_indicador_6B}
\alias{calc_indicador_6B}
\title{Calcula o indicador 6B: "Percentual de estudantes que permanecem pelo menos 7h em atividades escolares"}
\usage{
calc_indicador_6B(df_matricula, ano, verbose = TRUE)
}
\description{
Calcula o indicador 6B: "Percentual de estudantes que permanecem pelo menos 7h em atividades escolares"
}
