% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_8.R
\name{calc_indicador_8A_anual}
\alias{calc_indicador_8A_anual}
\title{Calcula o indicador 8A: "Escolaridade média da população de 18 a 29 anos"}
\usage{
calc_indicador_8A_anual(df_anual, ano, peso = FALSE, verbose = TRUE)
}
\description{
Calcula o indicador 8A: "Escolaridade média da população de 18 a 29 anos"
}
