% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_8.R
\name{calc_indicador_8B}
\alias{calc_indicador_8B}
\title{Calcula o indicador 8B: "Escolaridade média da população de 18 a 29 anos de idade residentes em áreas com habitantes de menor escolaridade"}
\usage{
calc_indicador_8B(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 8B
}
\description{
Calcula o indicador 8B: "Escolaridade média da população de 18 a 29 anos de idade residentes em áreas com habitantes de menor escolaridade"
}
