% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_8.R
\name{calc_indicador_8D}
\alias{calc_indicador_8D}
\title{Calcula o indicador 8D: "Razão entre a escolaridade média da população negra e da população não negra de 18 a 29 anos"}
\usage{
calc_indicador_8D(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 8D em porcentagem
}
\description{
Calcula o indicador 8D: "Razão entre a escolaridade média da população negra e da população não negra de 18 a 29 anos"
}
