% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_8.R
\name{calc_indicador_8D_anual}
\alias{calc_indicador_8D_anual}
\title{Calcula o indicador 8D: "Razão entre a escolaridade média da população negra e da população não negra de 18 a 29 anos"}
\usage{
calc_indicador_8D_anual(df_anual, peso = FALSE, verbose = TRUE)
}
\description{
Calcula o indicador 8D: "Razão entre a escolaridade média da população negra e da população não negra de 18 a 29 anos"
}
