% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_9.R
\name{calc_indicador_9A}
\alias{calc_indicador_9A}
\title{Calcula o indicador 9A: "Taxa de alfabetização da população de 15 anos ou mais de idade"}
\usage{
calc_indicador_9A(df, verbose = TRUE)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}

\item{verbose}{exibe informações no console se True}
}
\value{
Indicador 9A em porcentagem
}
\description{
Calcula o indicador 9A: "Taxa de alfabetização da população de 15 anos ou mais de idade"
}
