% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_15.R
\name{checa_coerencia_disciplina_curso}
\alias{checa_coerencia_disciplina_curso}
\title{Verifica se a disciplina lecionada por um docente e a sua formação são coerentes}
\usage{
checa_coerencia_disciplina_curso(disc, codcurso)
}
\arguments{
\item{disc}{Disciplina lecionada por um docente}

\item{codcurso}{Código do curso da formação do docente (ver Anexo 5 dos microdados do censo escolar)}
}
\value{
TRUE se a disciplina e o código forem coerente; FALSE caso contrário
}
\description{
Verifica se a disciplina lecionada por um docente e a sua formação são coerentes
}
