% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_12.R
\name{calc_indicador_12B_anual}
\alias{calc_indicador_12B_anual}
\title{Calcula o indicador 12B: "Taxa de escolarização líquida ajustada na educação superior da população de 18 a 24 anos"}
\usage{
calc_indicador_12B_anual(df_anual, ano, peso = FALSE, verbose = TRUE)
}
\description{
Calcula o indicador 12B: "Taxa de escolarização líquida ajustada na educação superior da população de 18 a 24 anos"
}
