% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_15.R
\name{calc_indicador_15B}
\alias{calc_indicador_15B}
\title{Calcula o indicador 15B: "Percentual de docentes e de profissionais de apoio à docência da educação básica com formação específica de nível superior, obtida em curso de licenciatura na área de conhecimento em que atuam"}
\usage{
calc_indicador_15B(df_docente)
}
\arguments{
\item{df_docente}{DataFrame com dados carregados da tabela "docentes" do Censo da Educação Básica}
}
\value{
Indicador 15B em porcentagem
}
\description{
Calcula o indicador 15B: "Percentual de docentes e de profissionais de apoio à docência da educação básica com formação específica de nível superior, obtida em curso de licenciatura na área de conhecimento em que atuam"
}
