% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_17.R
\name{calc_indicador_17_anual}
\alias{calc_indicador_17_anual}
\title{Calcula o indicador 17: "Razão entre salários dos professores da educação básica, na rede pública (não federal), e não professores, com escolaridade equivalente"}
\usage{
calc_indicador_17_anual(df_anual, peso = FALSE, verbose = TRUE)
}
\description{
Calcula o indicador 17: "Razão entre salários dos professores da educação básica, na rede pública (não federal), e não professores, com escolaridade equivalente"
}
