% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_1.R
\name{calc_indicador_1A_censo}
\alias{calc_indicador_1A_censo}
\title{Calcula o indicador 1A: "Percentual de crianças de 4 a 5 anos na pre-escola"}
\usage{
calc_indicador_1A_censo(df_mat, df_pop, verbose = TRUE)
}
\arguments{
\item{df_mat}{DataFrame com dados carregados da PNAD Contínua trimestral}

\item{df_pop}{DataFrame com dados carregados com a previsão da populacao de Recife}

\item{verbose}{exibe informacoes no console se True}
}
\value{
Indicador 1A em porcentagem
}
\description{
Calcula o indicador 1A: "Percentual de crianças de 4 a 5 anos na pre-escola"
}
