% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_1.R
\name{calc_indicador_1B_censo}
\alias{calc_indicador_1B_censo}
\title{Calcula o indicador 1B: "Percentual da população de 0 a 3 anos que frequenta a escola"}
\usage{
calc_indicador_1B_censo(df_mat, df_pop, verbose = TRUE)
}
\arguments{
\item{df_mat}{DataFrame com dados carregados da PNAD Contínua trimestral}

\item{df_pop}{DataFrame com dados carregados com a previsão da populacao de Recife}

\item{verbose}{exibe informacoes no console se True}
}
\value{
Indicador 1B em porcentagem
}
\description{
Calcula o indicador 1B: "Percentual da população de 0 a 3 anos que frequenta a escola"
}
