% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_4.R
\name{calc_indicador_4}
\alias{calc_indicador_4}
\title{Calcula o indicador 4: "Percentual da população de 4 a 17 anos com deficiência que frequenta a escola"}
\usage{
calc_indicador_4(df, ano, peso = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{tabela matrícula do censo demográfico}

\item{ano}{ano do censo fornecido}

\item{peso}{retornar indicador ponderado?}

\item{verbose}{be communicative?}
}
\description{
Calcula o indicador 4: "Percentual da população de 4 a 17 anos com deficiência que frequenta a escola"
}
