% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_5.R
\name{calc_indicador_5A}
\alias{calc_indicador_5A}
\title{Calcula o indicador 5A: "Taxa de estudantes com proficiencia insuficiente em Leitura"}
\usage{
calc_indicador_5A(df, verbose = TRUE)
}
\arguments{
\item{df}{DataFrame com dados carregados da Avaliacao Nacional da Alfabetizacao}

\item{verbose}{exibe informacoes no console se True}
}
\value{
Indicador 5A em porcentagem
}
\description{
Calcula o indicador 5A: "Taxa de estudantes com proficiencia insuficiente em Leitura"
}
