% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_5.R
\name{calc_indicador_5C}
\alias{calc_indicador_5C}
\title{Calcula o indicador 5C: "Taxa de estudantes com proficiencia insuficiente em Matematica"}
\usage{
calc_indicador_5C(df, verbose = TRUE)
}
\arguments{
\item{df}{DataFrame com dados carregados da Avaliacao Nacional da Alfabetizacao}

\item{verbose}{exibe informacoes no console se True}
}
\value{
Indicador 5C em porcentagem
}
\description{
Calcula o indicador 5C: "Taxa de estudantes com proficiencia insuficiente em Matematica"
}
