% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_5.R
\name{calc_indicador_5C}
\alias{calc_indicador_5C}
\title{Calcula o indicador 5C: "Proficiência dos alunos do 3o ano do ensino fundamental em Matemática"}
\usage{
calc_indicador_5C(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da Avaliação Nacional da Alfabetização}
}
\value{
Indicador 5C em porcentagem
}
\description{
Calcula o indicador 5C: "Proficiência dos alunos do 3o ano do ensino fundamental em Matemática"
}
