% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_6.R
\name{calc_indicador_6B}
\alias{calc_indicador_6B}
\title{Calcula o indicador 6B: "Percentual de estudantes que permanecem pelo menos 7h em atividades escolares"}
\usage{
calc_indicador_6B(df_matricula, verbose = TRUE)
}
\arguments{
\item{df_matricula}{DataFrame com dados carregados do Censo Escolar}

\item{verbose}{exibe informações no console se True}
}
\value{
Indicador 6B em porcentagem
}
\description{
Calcula o indicador 6B: "Percentual de estudantes que permanecem pelo menos 7h em atividades escolares"
}
