% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_8.R
\name{calc_indicador_8A}
\alias{calc_indicador_8A}
\title{Calcula o indicador 8A: "Escolaridade média da população de 18 a 29 anos"}
\usage{
calc_indicador_8A(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 8A
}
\description{
Calcula o indicador 8A: "Escolaridade média da população de 18 a 29 anos"
}
