% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_9.R
\name{calc_indicador_9A}
\alias{calc_indicador_9A}
\title{Calcula o indicador 9A: "Escolaridade média da população de 18 a 29 anos"}
\usage{
calc_indicador_9A(df)
}
\arguments{
\item{df}{DataFrame com dados carregados da PNAD Contínua trimestral}
}
\value{
Indicador 9A em porcentagem
}
\description{
Calcula o indicador 9A: "Escolaridade média da população de 18 a 29 anos"
}
