% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_9.R
\name{calc_indicador_9A_anual}
\alias{calc_indicador_9A_anual}
\title{Calcula o indicador 9A: "Taxa de alfabetização da população de 15 anos ou mais de idade"}
\usage{
calc_indicador_9A_anual(df_anual, ano, peso = FALSE, verbose = TRUE)
}
\description{
Calcula o indicador 9A: "Taxa de alfabetização da população de 15 anos ou mais de idade"
}
