% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores_9.R
\name{calc_indicador_9B_anual}
\alias{calc_indicador_9B_anual}
\title{Calcula o indicador 9B: "Taxa de analfabetismo funcional da população de 15 anos ou mais de idade"}
\usage{
calc_indicador_9B_anual(df_anual, peso = FALSE, verbose = TRUE)
}
\description{
Calcula o indicador 9B: "Taxa de analfabetismo funcional da população de 15 anos ou mais de idade"
}
